agent/requirements.txt
requests>=2.28
PyYAML>=6.0
numpy>=1.22
Pillow>=9.0
opencv-python-headless>=4.5
picamera2>=0.3.12   # RPi CSI kamera – volitelné, fallback na opencv
pytesseract>=0.3.10
Flask>=2.3
markupsafe>=2.0
